+++ packages/ai/types.go
@@ -8,7 +8,7 @@ import (
 	"time"
 )
 
-const Version = "0.2.0"
+const Version = "0.2.2"
 
 type ProviderOptions map[string]any
 type ProviderMetadata map[string]any